backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pillow==11.1.0
python-multipart==0.0.9
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 